requirements.txt
discord
asyncio
ffmpeg-python
PyNaCl
python-dotenv
SQLAlchemy
google-api-python-client
requests
pytz